requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
tensorflow>=2.13.0
numpy==1.26.4
scikit-learn==1.3.2
pandas==2.1.3
python-dotenv==1.0.0